+++ app/src/main/java/com/cajetan/youtubeplus/data/VideoData.java
@@ -3,12 +3,23 @@ package com.cajetan.youtubeplus.data;
 import android.arch.persistence.room.ColumnInfo;
 import android.arch.persistence.room.Entity;
 import android.arch.persistence.room.PrimaryKey;
+import android.support.annotation.NonNull;
 
-@Entity
+import javax.annotation.Nonnegative;
+
+@Entity(tableName = "video_data_table")
 public class VideoData {
     @PrimaryKey
-    public int vIndex;
-
+    @NonNull
     @ColumnInfo(name = "video_id")
-    public String videoId;
+    String videoId;
+
+    public VideoData(@NonNull String videoId) {
+        this.videoId = videoId;
+    }
+
+    @NonNull
+    public String getVideoId() {
+        return videoId;
+    }
 }