+++ app/src/main/java/com/cajetan/youtubeplus/data/VideoDataDao.java
@@ -1,18 +1,21 @@
 package com.cajetan.youtubeplus.data;
 
+import android.arch.lifecycle.LiveData;
 import android.arch.persistence.room.Dao;
 import android.arch.persistence.room.Delete;
 import android.arch.persistence.room.Insert;
+import android.arch.persistence.room.OnConflictStrategy;
 import android.arch.persistence.room.Query;
 
 import java.util.List;
 
 @Dao
 public interface VideoDataDao {
-//    @Query("SELECT * FROM video_data")
-//    List<VideoData> getAll();
 
-    @Insert
+    @Query("SELECT * FROM video_data_table")
+    LiveData<List<VideoData>> getAll();
+
+    @Insert(onConflict = OnConflictStrategy.IGNORE)
     void insertAll(VideoData... users);
 
     @Delete